Register components automatically in ECM
When developing the game, it's a hassle to have to manually register each new
component type. Now we allow unregistered components to be added to entities and
be registered automatically.

This brings in some performance issues (setting a component could take a linear
time instead of a constant depending on whether it's registered or not), but
it's all right during the development phase and once things settled, the user
can register all the components in one place.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>