Strictly separate turns
This removes some of the hackiness of the `update` procedure. We've separated
the systems that should only be run within the context of a game turn from the
rest of the update code -- rendering, game exit/restart key processing, etc.

Now we explicitly run all the systems for the Player entity. If the player took
an action, we run the same set of systems in the same order for all the NPCs.

Right now, this means duplicating the systems. The code should be refactored.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>