Refactor everything around the AP system
This removes most of the hackiness we've had to put into the code to deal with
the fact that players moves first, then the NPCs, but only if player actually
did anything.

We're now using a generic turn-based system. Each participating entity can be
either active or inactive. Active entities have a certain ammount of action
points to spend. Once they've spent them, the other party becomes active and
gets APs to spend. Then a new turn begins.

To properly test this, the player now has 3 action points per turn and the NPCs
have 2 APs each. The player moves first, then the entities.

We've also made the system code more defensive: if the entity being processed by
a system doesn't have all the components the system needs, it won't be
processed. This poses a performance penalty, but it's not so bad as to affect
the gameplay and we've had a lot of bugs without the checks.

We will probably want to automate the system processing and checks later on:

    @system(Position, MoveDestination, Turn)
    def movement_system(e, pos, dest, turn, ecm):
        pass  # `e` is guaranteed to have Position, MoveDestination and Turn

    ecm.run_system(movement_system)

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>