Enable the addiction effects again
The State of Mind is now going down at the end of each turn. It now has its own
system that tracks whether it's time for the next turn or not.

I *think* it may eventually make more sense to track turns outside of the
entities and components (since that's a global state much like the time elapsed)
and to pass a turn-based `dt` value to the system that need it, but that can wait.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>