Add field of view and exploration
The unexplored territory is now black, are visible only within PCs line of sight
and a close distance.

We had to remove the default empty background console and add the empty tiles
back. The reason being, we have to track whether they've been visited or not.
That in turn meant readjusting the fading factor on low state of mind.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>