Make sure the initial dose is always visible
If it happened to be generated in one of the corners farthest away from the
player, it was invisible because of the circularity of our field of view.

So now we generate a larger explored area which is outside of FoV and make sure
the dose is counted as explored, too.

Incidentally, this seems to be a good way to handle the exploration itself: to
have 2 circles: a smaller one for field of view and a slightly larger one for
explored-but-not-immediately-visible areas.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>