Lay the ground for the cheat mode
We no longer show the State of Mind or Tolerance values by default. The code is
still there, though, and will be displayed when we set `CHEATING` to `True`.

There's no way to set that in-game yet, but there will be in the future. We'll
also show the entire map and make the player immortal. This should make testing
certain game features (especially cumulative effects) more straightforward.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>