Add helper drawing functions with proper layering
This cleans up the code dealing with the background issues with tcod console
blitting.

We were using multiple consoles to handle different rendering layers (such that
no matter what gets drawn when, monsters are always on top of the background,
etc.) and we set a blit-transparent background colour that ensured empty cells
of a higher layer don't hide glyphs on a lower layer.

However, the blit-transparency was interfering with us wanting to set a
background colour of a cell independently on what glyph was or was not rendered
there. If we had a cell with a glyph and a foreground colour set but having a
transparent background, it would not be rendered because tcod's blitting only
looks at the background colour, not at the cell's contents.

So this adds two helper functions: set_background and draw_char. The former sets
the background colour of the cell and the latter sets the glyph and the
foreground colour of the cell. They're independent of each other: no matter the
order or the values, these two will never interfere with each other.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>