Fix the teleporting bug.
When the PC was within the irresistible dose range, but the pathfinding caused
them to bump into a monster, the TCOD path was walked even though the player
didn't actually move (they attacked the monster standing in the way).

This meant that the next turn the player teleported to the next path step. This
fix makes sure we walk the TCOD path only once the move has happened.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>