Rename AI.kind to AI.state
The 'aggressive'/'idle' thing is really the value in the AI's state machine so
we name it as such. Kind on the other hand refers to the actual behaviour of the
AI. So far we have only one type: charge when you see the player.

The idea is to add a hunting pack behaviour that signals to the other members of
the pack, too.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>