Fix for the walkability issue
This should work with multiple entities running about and be more performant,
but the downside is we have to be extra careful with changing entity's position.
Whenever a position is changed, we must also set the correct solidity values for
both the previous and the new cell.

It should be fine since the position should only be modified in the
`movement_system`, but since there is no enforcement right now, it's not ideal.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>