Update Makefile to generate a launcher
`make build` now creates the game binary in `bin/dose-response` and a
launcher in `./dose-response`. That means we can now run the game
without having to specify LD_LIBRARY_PATH.

This is extra useful when we want to run a replay different than the
last one (i.e. where `make replay` wouldn't help).

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>