Switch from entity iterators to ID-based access
If we ever need to access components to more than one entity, we can't
iterate over them because that gets us into the borrowed pointer mess.

So what we do instead is: iterate over the entity ids, and access them
in the systems as needed by passing the ECM in.

This makes the code more complicated, but it's the only way how to
implement bump and combat systems without significantly altering the
EntityManager implementation.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>