Move the AI system to a separate module
We'll probably want to move different systems to separate modules eventually.
Some of them are going to require specific imports or helper functions that
would polute the namespace for other systems.

This is certainly true for the AI system so let's start there by putting it in a
separate module.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>