Hack around the segfault bug
So it's as I thought -- the `PathData` in `Map::find_path` is created on the
stack and passing it to the `cb` function can cause a segfault. It's just hadn't
happened until now.

Valgrind doesn't report anything so I guess either Rust or LLVM shields us from
that (or maybe it's the optimization switches or something?).

Anyway, so I've added vector of large enough capacity not to get resized anytime
soon and I put the `PathData` instances there. This means I can (reasonably)
safely point to them and pass that pointer to the path-finding callback.

But it's a messy hack. We need to come up with a clean solution that doesn't
have memory issues.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>