Make the bloody turn system finally robust
Jesus. That wasn't so hard, was it? So, we have a global turn counter and we
keep track of the current side. Each tick we process all the active
entities (i.e. whose side is equal to the current side). They can spent from
zero to max of their action points.

Once all the current entities have 0 AP, the turn system switches to the next
side and sets the action points to all the entities of the new side to their
max.

If this means switching to the first side (i.e. all entities spent all their
action points this turn), we increment the turn counter as well.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>