Reorg the initial world generation
We need things to be more modular in order to support things such as restarting
the game or moving to a new (randomly generated) location.

This splits populating the EntityManager with the generated entities, generating
the player entity, generating a completely new game state and preparing a game
state for replay into separate functions.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>