Fix the broken replay when stunned
We didn't get our turn stuff *quite* right before.

When the player was supposed to leave the stun, what was supposed to happen was:

-- tick --
1. New turn
2. Remove the stun effect
3. Process player's input
-- end of tick --

Here's what happened instead:

-- tick --
1. New turn
2. Process player's input (there likely wasn't one since ticks are fast and I am
slow)
3. Remove the stun effect
-- end of tick --

-- next tick --
1. Process player's input (a key was pressed)
2. Move to the specified location
-- end of tick --

This broke on update, where we have all the keys buffered so this is what
happened instead:

-- tick --
1. New turn
2. Process player's input (we read the next key from the queue)
3. Try to move but be blocked by the still-active stun effect
4. Remove the stun effect
-- end of tick --

This caused the timelines to diverge (in effect, the replay was stunned a turn longer).

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>