Fix the map migration bug
The EntityManager's `clear` wasn't implemented correctly and it was giving wrong
IDs which due to bugs in the game loop and `id_iter` gave the illusion that
everything was working fine.

This adds some tests and fixes the issues with Entity Manager and also fixes the
systems accordingly. Essentially: we can no longer rely on the Player ID being
0. It changes when we move to a new area.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>