Add codegen that builds components and the entity manager
The `build_ecm.py` script takes a file as an input. In that file are
newline-separated component definitions. The script reads them up and spits out
a file with the same components declared as pub struct or enum along with an
implementation of the entity manager that has access methods such as
`has_position`, `get_destination`, etc.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>