Add a Path component
This doesn't do anything yet and will require a rework of the Map helper. The
idea with components is to only use primitives. No owned pointers, destructors
or the like since that imposes borrowed-pointer-deferrencing and lifetime issues
that we don't want to deal with.

If this becomes a problem, we can address it later but the path was the only
problem so far.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>