Make colour fading independent of tile's colour
This makes things a bit more generic. When we're animating a colour fade, all we
use are the `FadeColor` and `ColorAnimation` components. Whenever an animation
is taking place, `Tile.color` is ignored. As soon as animation ends, both
components are removed and `Tile.color` wins again.

This way: the animation system is cleaner, we can set the final color of the
animation by just changing the tile colour when we set the animation up (useful
in the fade to death situation) *and* we can have the animation use completely
different colours from the one that the tile uses. Which should be helpful e.g.
in the High SoM animations.