Don't remove Position in kill_entity
A position should be able to persist even after an entity is dead (useful for
fadeout animations, showing corpses and the like).

This exposes the limitation with our Map implementation -- that it supports only
two entities in a single place at the same time. Since the dead entities are now
lying around (albeit invisible and in no way interacting with the world), the
map has a problem when another entity steps in. We should just remove the Map anyways.