Fixed the killed-by-invisible-monster bug
We got actually bitten by duck typing here! This must be the first.

The Anxiety entity was dead, but it had the `AttributesModifier` component on
and the `interaction` system was detective "interactive" entities incorrectly.

So when the player stepped on the invisible remnants of the monster, the
`interaction` system kicked in and removed 1 Will, killing the player.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>