Set the lib path in the generated ELF file
This allows us to drop the `LD_LIBRARY_PATH` env var to reference the
bundled libtcod*.so libraries.

Using the --rpath linker invocation, the additional `$APP_DIR/lib`
location is added to the generated ELF executable, meaning we can use
the exe directly instead of having to wrap it in a shell script.

This is FUCKING AWESOME.