Fix the bug where enemies step on player's corpse
When the player was surrounded by monsters and killed by one, one of the
monsters occasionally moved into the player's place.

This is because the `kill_entity` function removed the `Solid` component.

We've renamed the `DeathTile` component to `Corpse` and added the `solid`
attribute. If a corpse is solid, `kill_entity` will no longer remove the
solidity.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>