Proof of concept implementation of eating
Um, so this is just directly called from the InputSystem because guess what, a
game designed completely around the concept of bumping into things has trouble
understanding a non-movement action.

So:

* Eating causes the entity to wait even if they don't have any food because we
  have to set a dummy Destination value
* The Eating system isn't included in the list of systems to be ticked because
  we don't have anything that would trigger it (such as the Destination for
  movement)
* Panic and Stun ignore the Eating system because they only work on a
  Destination level

Jesus.

But it still kinda works and I like it. I think instead of having the player
start with some foods in their bag, we should just scatter 1 - 3 around the
starting area.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>