Handle the food properly in Input system
This removes the hack where we've always relied on the Input system to set a
Destination (which makes no sense). The system should set any components that
result from the specified user action (in this case either movement or eating).

As a result, this also fixes the bug where we spent an AP while eating even
when the player didn't actually have any food. This was because we had to set
their destination to a dummy value (the current position) which got interpreted
as waiting.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>