Produce the binary with cargo-lite
Had to update the toml lib from the source to fix the bug with a coma inside of
a list of strings (FFFFFUUUUUUUUUUUUUUU). And update cargo-lite to allow setting
`--out-dir` via `rustc_args` so it can put binaries outside of `src` as well.

With that in place, we can now get the binary by running `cargo-lite build`.