Move initialising the world to its own procedure
This is the first step towards fixing the broken restart on F5.

Problem is that the restart code was written prior to having a concept of
`emhyr::World` and when we add the systems in `main` they don't get added again
in `update`. Plus there's been some code duplication there anyway.

So we've moved that sort of stuff to a standalone procedure, which will allow us
to call it both from `main` and from `update`.