More changes to our Cargo/Makefile setup
Since cargo and make both check for the freshness of files and cargo
doesn't have a way to do a force rebuild, we just always shell out to
Cargo from make (i.e. no messing with checking the targets etc.).

Cargo's `build` directive makes sure the libtcod.so symlinks are always
present in the `deps` directory so the compilation succeeds.

Unfortunately, Cargo's unable (afaik) to set up the rpath for the
binaries it spits out, nor can it run a post-compile job so we have to
set the executables' RPATH in the Makefile.