Move monsters out of Level to a Vec in GameState
This gives us a fixed ordering and allows us to do a lot of operations on
monsters directly. Level still keeps a record of each monster's position for
fast position-based lookup, but it stores a monster's ID instead of the actual
struct now.

This should fix the replay inconsistency, but that hasn't been tested yet.