Don't hang the game if monster can't find path
We forgot to handle that particular code branch in monster action/movement code.

The player was behind a wall or whatever but close enough for the monster to see
it. It tried to chase them, but because of the blockage, path-finding / walk
step returned None. And we ignored that value.