Enable additional fields in Item
Similar to what we did with the Monster: the previous `Item` enum has been
renamed to `Kind` and became a `kind` field in the new `Item` struct.

This will let us do things such as attaching a `OnUseEffect` struct which could
contain a bunch of modifiers such as `state_of_mind`, `dose_tolerance`, etc.