Make the status panel vertical
We want to make the displayed map square to in order to make map-moving
animations more regular. And because of the monitor shapes, it makes
more sense to make the panel vertical and move it to the side.

So it's more like 868-HACK now.

This has the added benefit that each status item is on its own line.
This makes them more readable as well as easier to code.