Fix the ranges for food and dose explosions
Turns out we were using a range that calculates tiles which fall within
a (as in non-integer) real circle (that we use for exploration radius
etc.) to the explosion effects.

But those we do actually want to be squares. So if it's a radius of 2,
it will always save you from the Depression.

So we've added a new iterator over the square area (called `SquareArea`)
and we've renamed the previous one to `CircularArea`.