Fix the crash
When you get too close to the left edge of the screen and are
intoxicated enough, the field of view can extend beyond the area on the
screen. In that case, rendering the field around the irresistible dose
can cause an assertion error.

We've fixed this by just never calling any TCOD rendering code or
asserts in our engine when out of bounds. This is not the first crash
like that and if we try to write a few out-of-bounds tiles than so be
it.