Fix the point dimension comparison
We had to implement all the comparison operators directly, because we
treat `less than or equal` as `both components are less than or equal`,
but that doesn't mean that these operations can be split.

E.g.:

    Point(0, 0) <= Point(1, 0)

But these are both false:

    Point(0, 0) < Point(1, 0)
    Point(0, 0) = Point(1, 0)

Though the conventional understanding here would be that either one of
them must be true.

The current behaviour, although confusing is great for bounds checking
which is why it's there. It should probably not be used for anything
else because of the confusion this can raise.