Disable 32-bit Linux & Mac and no more cargo run
So the 32-bit cross compilation is failing for Linux and MAC so let's
disable those for now. Same goes for MinGW builds on Windows.

And as far as I can tell, the rest seems to be working, but the CI is
failing because `cargo run` can't succeed in the environment (because we
create an OpenGL window etc.)