Improve the mind progress bar
We've updated the bar width so that for small ranges (e.g. withdrawal),
the points on the bar correspond with the values 1 to 1 (so e.g. when
you have 1 ponint left before you die of exhaustion, that's obvious from
the progress bar).

We've also placed it in the right height (right below the mindstate
name).