Doses can be picked up on maximum Will
When the player reaches the maximum Will (currently 10), they can pick
up the doses. We now show food and doses separately in the sidebar and
each has a numeric hotkey associated.

Pressing `1` eats food (if any) and `2` uses a does the player is
carrying. We need to differentiate between the normal and strong doses,
though.