Add convenience methods to RangedInt
We now have is_min/is_max checks as well as methods that set the current
value to the minimum/maximum of the range.

This makes the code shorter and easier to read.

Turns out, there's no longer a need for the `set` method anymore, so
we've commented it out for now.