Introduce World and Chunks
Instead of having a single level that's as big as the world, we split
the world into chunks of smaller size -- each with its own random
generator (but all determined from the original) and level.

This doesn't compile yet, I've mostly just written the code that's
necessary to add the chunks, but nothing more.