Implement World::cell
It's relying on two more `unimplemented` invocations, but the interface
seems good now.

Problem is, the world now needs to be mutable even for seemingly
immutable operations (because we may need to generate a chunk in). We
could return Options everywhere or panic in a "page fault", but let's
try this approach and see where it goes.

Had to update some of the code relying on this mutability change.