Implement the cell iteration kind of somewhat
Basically mutable iteration is now a no-op and the immutable one takes a
closure (returning an iterator over possibly nonexistent chunks isn't
very easy).

Probably buggy as hell, but it's theoretically the biggest hurdle and we
should be able to handle the rest and figure out any additional
restructuring that may be necessary.