Add a bunch of notes and monster-related code
This is basically me trying to figure out how to do chunked monsters.
I've got an idea now and working towards it but it's not even compiling
atm.

Basically, since the monsters need to be chunked, we'll make them part
of the World struct and we'll need to be careful to avoid double
borrows.