Add commandline options for choosing the backends
We've added clap as a dependency, because the options are now more
complex so let's use something designed for the job.

It all works now \o/.

I guess the next step is to make each backend into a Cargo feature and
compile it conditionally, but this is already quite neat.