Fix the replay inconsistency bug
It was because of a timing issue. When the screen re-centered itself,
the player usually didn't do much in that time, but during replay,
several turns may have elapsed. This caused that any new chunks that
were created happened at different turns and that means the monsters
that were created with them happened at different turns as well and
therefore behaved differently.

We're now blocking command/player processing during the screen
recentering animation, too.