Move the explosion effect into the animation
The render and covered_tiles methods have been joined into one which
returns the covered tile's position, explosion colour and the effect
that should happen under the animation.

The last bit is a KILL (the blue background for killing monsters) or
SHATTER (red background for shattering the walls) and it's a bitflag so
you can have all, none or either.