Let smaller doses be pickable sooner
Previously, the player had to have the maximum value of the Will. But
since they have different irresistible areas, it makes sense that the
smaller doses can be carried around earlier than the stronger ones.

This has probably made the game way easier, so we'll have to re-balance
that later.