Don't generate verifications in the release builds
The replay verifications as they currently stand take up massive amounts
of space (20MB for a playthrough) so we're disabling them in the release
builds.

With the judicious use of the "fuzzing", we should be able to ensure
that there are no desync issues in the releases, so the replays should
be fine without the verifications.