Add Behavior to monsters/npcs
Apart from the AI state which is what the mob's currently doing, we're
adding a new variable called `Behavior`. This describes the general
behavior of the monster. So e.g. a lone attacker vs. a pack of wolves
vs. a friendly.

They can share some of the states, but have different state machines.