Implement fullscreen
This time it actually works on linux \o/.

Turns out, the window rebuild thing from glium still isn't
implemented, but we're able to get the actual glutin window and call
`set_fulscreen` on that. And that just works.

So: Alt+Enter to switch fullscreen. Works with multiple monitors. And
no weird resolution switches either.

We'll have to update the rendering to fix the different aspect ration
-- but we need to do that anyway for when someone resizes the window.