Change the priority of of the 1.0 roadmap items
Basically, if we want to ever finish this, we'll have to reduce the
scope. What everybody keeps saying and what I thought I was doing but
we need to do more.

The most important things are: NPCs, proper ending, game balance and
then sort of polish essentials: save/load, fullscreen, nicer colours.