Pass the data correctly
Turns out, we were getting the clientside pointer to the memory a bit
too early -- once it grew, the old memory got zeroed and the data
moved somewhere else.

So this is basically a cleanup plus demonstration of creating a new
vec of "drawcalls", reading it with javascript all while keeping the
game state around.

No statics are necessary, either.