Isolate the Dose Response JS into one function
That way we can control its execution say by showing a "click here to
play" banner or whatever.

We also don't pollute the global namespace as much and we take the
path to the wasm file as a parameter now which means it can be driven
externally as well.

As a benefit to this, releasing a new build on the project website is
just a matter of copying the appropriate *.js and *.css files over.