Postpone using the font in wasm/context
For some reason on some browsers (my Firefox on Windows 7) the "click
to play/loading..." text doesn't show up because the font is not
loaded or something?

I tried putting everything under window.onload but that didn't work
either. Anyway, using the font only after the wasm module has finished
loading seems to work everywhere so that's what we'll do.