Process screen fading animation outside of windows
The `process_game` code should be able to push the window any time it
wants, not having to wait for the fading animation to finish. And not
having to do it there either.

So the screen fading is now being done in the main update function.

Sadly, as a result, we now display the endgame screen and uncover the
window during the fadeout not fadein, which is not as pretty.