Only mess with the Canvas after wasm is loaded
The caller might want to do something like show a "loading..." text on
the canvas before the game starts. This is impossible if we start
changing its size, font etc. as soon as `play_game` is called.

So let's defer it to the time we actually have everything we need to
play the game.