Use MessagePack for JS drawcalls
This adds the MessagePack serialization library, uses it to serialize
the wasm drawcalls, load those from JS and then implement each
drawcall correctly.

This means we now render `Text` differently from a glyph (the letters
are no longer aligned to the grid so they're much more readable), but
it means that our current layouting is broken, because that's all done
game-side and expects grid-aligned text.