Fix the drawcall sorting
This is described in the renderer in much greater detail, but
basically: we have to drop the drawcall sorting because it's not
really workable.

Instead, we're doing a multi-phase pass of the drawcalls that should
handle all the rendering confusion for us.

It's possible that we will have to turn this into another stream of
drawcalls for wasm, but at least for now, this does actually work.