Create a main menu struct
This is exploratory coding. We need to have a way to read the
highlighted option in the main menu that's available both from the
game code (without rendering) and the render code (with rendering).

So this adds a struct which will lay everything out and return the
option under the mouse. And this can be used in both cases.

This doesn't compile yet, but it looks like a decent approach so now
we just need to fill in the gaps.