Fix the fullscreen scaling issue
Instead of calculating the render area based on the width/height
ratios, we calculate the highest possible size of a tile and make sure
it's a whole number.

That way the math gets easier and more straightforward. And hey, it
fixes the fullscreen mouse position issues because that was caused by
accumulation of the fractional part.