Up the dose tolerance increase and strength
This increases the tolerance of all doses and increases the strength
of the stronger ones.

Which means that the player will get used to the normal doses more
quickly, but the stronger ones should still have some kick then.

Obviously we need to play a lot to see whether this actually feels right.