Save validations even on optimised builds
Now that we can differentiate between an optimised (`--release`) and
ready-for-distribution build, the validations should be added even on
the optimised builds.

I playtest the game increasingly with the optimised builds to
experience the performance as it should be but I still need access to
the replays and be sure they're correct.