Remove SDL drawcalls
The `engine` module now has a new `Drawcall` struct which will the new
rendering interface. It is modeled after the SDL calls: the only
operations are creating a rectangle and blitting a texture.

Anything else (including text layout) should happen in the "userspace"
although it can still live in the engine module.

This will make our renderers simpler.