Update to glutin 0.15.0 and winit 0.14.0
This is mostly a testing update to see if the latest versions work for
us. I'm using my github forks for glutin and glium here.

There's probably a way to do this with cargo patch or replace sections
or something but I can't be bothered to figure that out now.