Add logging
This adds the basic logging facility as well as a few usage examples
in `main.rs`. We now log both to the console and to a file.

Obviously the `simplelog` crate (which provides the logging backends)
doesn't compile on wasm, so it's not used there. But the `log` crate
can be, which means that we could potentially write a backend that
writes to the browser console or something.