Fix the shaders on OSX
Apparently, our fragment shaders were incorrect: the usage of
`texture2D` and `gl_FragColor` did not fit the selected profile.

This seems to be fine on Windows and Linux, but the shaders failed to
compile on the Mac.

This should use the correct syntax as well as update the webgl
transpiler to produce the expected output.