Use 7zip to create the windows tarball
Tar wasn't working right (couldn't figure out what was wrong from the
appveyor log but maybe tar isn't actually on the system).

Turns out, 7z is able to create a tarball, but it can only gz a single
file, so you need to create it the old-fashioned way in two steps:
first tar it up and then gzip it.