Add script that builds the final releases
It downloads the CI artifacts from github and does any post-processing
that's necessary. Things such as including the dependencies in the
right places and so on.

The output of this is created in `target/publish/version` and is
intended for distribution.

We'll probably want to move the note etc. generation in there as well
and just ship the "naked" binary in the github releases.