Fix the dose order from weakest to strongest
In the UI, the strong dose appeared right after the normal dose before
the cardinal and diagonal doses.

That ordering was incorrect and it did not correspond to the keys
assigned to each dose.

So this fixes the ordering (which comes from enum `Kind::iter`) as
well as how we order the items in the code (e.g. in the enum
definition) -- even though that has no outside effect.