Fix screen edge artefacts when scrolling
When the game scrolled to recentre itself, if we were on a fullscreen,
an extra row or column of the map would be visible.

This was because while we were filtering glyphs that didn't fit the
screen, we rendered anything that was even partially visible in full.

This has added checks that ensure that only the content that should
actually be visible on the screen gets rendered -- even if it means
only a portion of a glyph/tile.