Fix panic on replacing Victory NPC
When we created a Victory NPC when there already was one, we first
created the new one and then tried to delete the old one.

This caused a panic because of how we're storing them in the level
hashmap. The storage issue should be addressed properly but for now,
we've just corrected the action order: first delete the old one then
add the new one.