Show all tiles under display
When we're scrolling towards the Victory NPC, a lot of the path would
remain blank even though it should have been visible.

This updates the render function to show the area that's actually
under the camera, rather than what the update game thinks it is for
e.g. simulation purposes.

It feels a little choppier now, but I guess that's something we can
sort out later.