Try to read the git commit from travis/appveyor
Both Travis and Appveyor set environment variables containing the git
commit the code is being built from. Using it is preferable from
calling `git` directly (it may not even be installed on the builder).

So we try to read those variable first and only call git directly as a
fallback.