Fix wasm crash
The web version would panic when a new game was started. I believe
this was because we were changing the state pointer to something else
and then freeing the wrong place or something. There's unsafe code
there.

By making sure we overwrite the "old state" memory (rather than the
pointer) with the "new state" mem, things start working again.