Add a signpost when a Victory NPC disappears
If the player gets High while there is a VNPC in game, they would
disappear. That may get confusing, however.

So now, there is a new object left at the place showing a message to
the player when they bump into it. The message should be enough of a
hint to make the player realise they need to stay sober the whole
time.