Make backend features more descriptive
Instead of `sdl` and `opengl` (which doesn't make sense because
they're both opengl), the features are now called `sdl-backend` and
`glium-backend`.

The `-backend` suffix is necessary, because the feature names must not
clash with crate names. This is much clearer what the features are
about though.

And if we decide to try a pure glutin+gl (dropping glium), that can be
called `glutin-backend` or possibly `glutin-gl`.