Add glutin+gl backend stub
None of the actual code is in yet, but we've got a new
`glutin-backend` feature and it's hooked up to main and engine. It
will be used for a pure Rust + OpenGl backend (using glutin and winit
but not glium).

This will give us a better comparison with SDL and ultimately drop the
glium backend I think. It's much less actively maintained and now that
we use raw opengl calls anyway, there's not much reason to keep using
it even if I do like the API better.