Toggle window decorations on Glium fullscreen
Turns out, the "fullscreen is not really fullscreen on winit +
Wayland" issue is because winit keeps the window decorations on for
some reason. When we remove them before switching to fullscreen
explicitly, we get the real deal.

So we just always do that for now, but we also document this in the
list of winit issues.

That said, forcing the X11 backend is less of an issue with this.
The window decorations still look different though.