Remove individual tile's offset_px
Instead, make it a "global" value set on `engine::Display`. We don't
allow for any subpixel position (other than the "camera scroll"), so
this is unnecessary.

More to the point, it caused the scrolling glitches, because some of
the tiles (the "default" unset ones) have always set it to zero
whereas everything else set it to other values. This lead to multiple
tiles "rendering" on top of each other and (I believe) z-fighting.

Fuck z-fighting.