Enable CLI arguments conditionally
This changes our CLI/Clap config to only use an argument if the
corresponding feature has been enabled.

E.g. if the cargo binary did not have the "replay" feature set, we
won't show or respond to any of the "replay" commands (such as
"replay-full-speed").

This is less confusing to the users because they won't be exposed to
nonexisting arguments.