Remove glium
Glium is in a maintenance mode and doesn't seem to update its
dependencies very often (e.g. as of right now it's still using glutin
0.17 despite newer versions being up for over 6 months ago).

There still doesn't seem to be a clear maintainer, while Tomaka has
moved on to other things (he still occasionally merges patches, but
I'm not really confident in the future of the project).

It's a real shame, because I very much prefer glium's API to pure
opengl, but alas.

I'd love to help out, but given my time commitments (and my own
history with tcod-rs), I'd very likely be wayyy worse.

And since we're using "raw" opengl with the sdl backend, we actually
get to share some of that code (as well as having some confidence in
it).

I'm leaving the actual backend code in, but I've removed the
references from Cargo.toml, main.rs etc.

Tomaka, if you by some chance get to read this, thank you so much for
all your work in the Rust gamedev libraries! They've been tremendous
for Dose Response and despite not fulfilling your hopes, glium has
been excellent to me.