Use the glutin backend for dev by default
Glutin seems to be working well enough as the default backend -- for
dev at least.

This means as of now on, running Dose Response from a git checkout
will result in a pure Rust version. No need to install any external
libraries! \o/

The `prod` feature still builds SDL as well, but uses `glium` as the
default, too. We'll need to do a more thorough testing whether that's
what we actually want to do, but I hope it is.