Position glyphs vertically as well
This fixes the graphical glitches. The glyphs are now placed along the
top of the tilemap and they start at a new line when they reach the
end.

NOTE: having the tilemap an obscene 1024x1024 apparently slows the
graphics on load. It takes about a second before anything is shown on
the window.

That only happens in the debug build though. The release one is fast
enough.