Use pixel-based coordinates into the texture
They're now generated as such from the build script (instead of the
previously used "tile indices").

This doesn't have any significant advantages right now, but it will be
crucial when we start packing multiple glyph sizes in the same font
map. At that point, a simple "third column, second line" coordinate
system no longer works because the glyphs are not aligned to a
single grid.